[Feat] scala 2.12 support.
