+++ streamx-flink/streamx-flink-connector/streamx-flink-connector-influxdb/src/main/scala/com/streamxhub/streamx/flink/connector/influxdb/sink/InfluxDBOutputFormat.scala
@@ -21,7 +21,7 @@ package com.streamxhub.streamx.flink.connector.influxdb.sink
 
 import com.streamxhub.streamx.common.util.Logger
 import com.streamxhub.streamx.flink.connector.influxdb.bean.InfluxEntity
-import com.streamxhub.streamx.flink.connector.influxdb.function.InfluxDBFunction
+import com.streamxhub.streamx.flink.connector.influxdb.function.InfluxFunction
 import org.apache.flink.api.common.io.RichOutputFormat
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.configuration.Configuration
@@ -30,7 +30,7 @@ import java.util.Properties
 
 class InfluxDBOutputFormat[T: TypeInformation](implicit prop: Properties, endpoint: InfluxEntity[T]) extends RichOutputFormat[T] with Logger {
 
-  private val sinkFunction = new InfluxDBFunction[T](prop)
+  private val sinkFunction = new InfluxFunction[T](prop)
 
   private var configuration: Configuration = _
