+++ streamx-flink/streamx-flink-connector/streamx-flink-connector-redis/src/main/scala/com/streamxhub/streamx/flink/connector/redis/function/Redis2PCSinkFunction.scala
@@ -17,10 +17,10 @@
  * limitations under the License.
  */
 
-package com.streamxhub.streamx.flink.connector.redis.internal
+package com.streamxhub.streamx.flink.connector.redis.function
 
 import com.streamxhub.streamx.common.util.Logger
-import com.streamxhub.streamx.flink.connector.redis.bean.{RedisContainer, RedisMapper}
+import com.streamxhub.streamx.flink.connector.redis.bean.{RedisContainer, RedisMapper, RedisTransaction}
 import org.apache.flink.api.common.ExecutionConfig
 import org.apache.flink.api.common.typeutils.base.VoidSerializer
 import org.apache.flink.api.java.typeutils.runtime.kryo.KryoSerializer
